move duplicated like functionality into new `useLikeMutation` hook and add it to 2 pages where it was missing.
fixes #137
