+++ src/app/(auth)/(tabs)/index.tsx
@@ -13,8 +13,6 @@ import {
 } from '@tanstack/react-query'
 import {
   fetchHomeFeed,
-  likeStatus,
-  unlikeStatus,
   deleteStatusV1,
   postBookmark,
   getSelfAccount,
@@ -29,6 +27,7 @@ import CommentFeed from 'src/components/post/CommentFeed'
 import { useShareIntentContext } from 'expo-share-intent'
 import { useVideo } from 'src/hooks/useVideoProvider'
 import { useFocusEffect } from '@react-navigation/native'
+import { useLikeMutation } from 'src/hooks/mutations/useLikeMutation'
 
 export function ErrorBoundary(props: ErrorBoundaryProps) {
   return (
@@ -57,6 +56,8 @@ export default function HomeScreen() {
   const { hasShareIntent } = useShareIntentContext()
   const params = useLocalSearchParams()
   const [isPosting, setIsPosting] = useState(false)
+
+  const { handleLike } = useLikeMutation()
 
   useEffect(() => {
     if (hasShareIntent) {
@@ -204,30 +205,6 @@ export default function HomeScreen() {
       console.error('Error handled by share useMutation:', error)
     },
   })
-
-  const likeMutation = useMutation({
-    mutationFn: async (handleLike) => {
-      try {
-        return handleLike.type === 'like'
-          ? await likeStatus(handleLike)
-          : await unlikeStatus(handleLike)
-      } catch (error) {
-        console.error('Error within mutationFn:', error)
-        throw error
-      }
-    },
-    onError: (error) => {
-      console.error('Error handled by like useMutation:', error)
-    },
-  })
-
-  const handleLike = async (id, state) => {
-    try {
-      likeMutation.mutate({ type: state ? 'unlike' : 'like', id: id })
-    } catch (error) {
-      console.error('Error occurred during share:', error)
-    }
-  }
 
   const handleShowLikes = (id) => {
     bottomSheetModalRef.current?.close()
