+++ src/app/(auth)/profile/bookmarks/index.jsx
@@ -1,3 +1,4 @@
+//@ts-check
 import { Stack, useNavigation, useRouter } from 'expo-router'
 import { SafeAreaView } from 'react-native-safe-area-context'
 import { Text, View } from 'tamagui'
@@ -7,6 +8,7 @@ import { useInfiniteQuery, useMutation, useQueryClient } from '@tanstack/react-q
 import FeedPost from 'src/components/post/FeedPost'
 import { useCallback, useLayoutEffect } from 'react'
 import { Storage } from 'src/state/cache'
+import { useLikeMutation } from 'src/hooks/mutations/useLikeMutation'
 
 export default function LikesScreen() {
   const router = useRouter()
@@ -14,6 +16,7 @@ export default function LikesScreen() {
   useLayoutEffect(() => {
     navigation.setOptions({ title: 'My Bookmarks', headerBackTitle: 'Back' })
   }, [navigation])
+  const { handleLike } = useLikeMutation()
   const userJson = Storage.getString('user.profile')
   const user = JSON.parse(userJson)
   const renderItem = useCallback(
